P0-1/P0-2 MSVC 化: 复合字面量改写 + crt_shims 裁剪 + GCC 扩展清理 (syntax gate 绿)
- 复合字面量: (LARGE_INTEGER){...} x55 -> PECMD_LI(x), (FILETIME){0,0} -> PECMD_FT(0,0)
  (辅助构造函数加于 win32_stub.h / stubs_common.h 各一份, 语义与原形式等价)
- 标量->LARGE_INTEGER 非法强转 x22 -> PECMD_LI(...) (反编译器 cast 残留)
- 裸变参参数列表 (...)->() x21 (MSVC C 不接受纯 ... 参数列表)
- crt_shims.c P0-2: 190 个 CRT 机制桩/libc 同名桩删除(系统 CRT 接管, 调用点原型已核),
  保留 12 个 FUN_ math 缺口桩(core_b3_remaining.c 真实调用, 无系统对应)
- __int128/__uint128_t x4: mulhi 用可移植 PECMD_MulHi64(Hacker's Delight),
  (a*b)/100 用 PECMD_MulDivU64(逐位长除), 取低64位处直接普通乘法截断
- void* 算术 x7 -> (char*) 显式字节步长; g_pMapBlk+6 修正为 +0x18(Ghidra int* 步长,
  与相邻取字段代码一致)
- 标签后空语句 x2, WSTR 续行段丢 L 前缀 x1(\\x3000 越界, 合并单字面量)
- GetCurrentThreadId 返回类型对齐 DWORD; core_sys.c 补前置声明(C2371);
  core_b7c.c 补 string.h(隐式 memset 声明冲突); stubs_common.h GetPartitionLayoutEntry
  去 static(C2129); __debugbreak GCC 垫片限 _MSC_VER 外
- runners 四脚本 GBK 控制台兼容(stdout/stderr reconfigure utf-8)

验证: tools\\msvc_build.bat syntax 全绿 (237 error lines -> 0)

+++ win32_api_stubs.c
@@ -149,7 +149,7 @@ DWORD GetCurrentDirectoryW(DWORD n, WCHAR *buf) { (void)n;(void)buf; return 0; }
 uint64_t GetCurrentProcess(void) { return 0; }
 uint64_t GetCurrentProcessId(void) { return 0; }
 uint64_t GetCurrentThread(void) { return 0; }
-uint64_t GetCurrentThreadId(void) { return 0; }
+DWORD GetCurrentThreadId(void) { return 0; }
 int GetCursorPos(void *p) { (void)p; return 0; }
 HDC GetDC(void *h) { (void)h; return (HDC)0; }
 uint64_t GetDIBits(void) { return 0; }
